refactored reducer & lib functions
