+++ packages/components/acter/landing-page/members-section/index.tsx
@@ -7,13 +7,9 @@ import { DisplayActers } from '@acter/components/acter/landing-page/members-sect
 import { DisplayMembers } from '@acter/components/acter/landing-page/members-section/display-members'
 import { Selectors } from '@acter/components/acter/landing-page/members-section/selectors'
 import { LoadingSpinner } from '@acter/components/util/loading-spinner'
-import { getActerJoinSelectors } from '@acter/lib/acter/get-acter-join-selectors'
-import { mapFollowersByType } from '@acter/lib/acter/map-followers-by-type'
+import { getFollowersByType } from '@acter/lib/acter/get-followers-by-type'
 import { useActer } from '@acter/lib/acter/use-acter'
 import { MemberType } from '@acter/lib/constants'
-import { ActerConnectionRole } from '@acter/schema'
-
-const { ADMIN, MEMBER } = ActerConnectionRole
 
 const { ACTERS, PEOPLE } = MemberType
 
@@ -26,18 +22,11 @@ export const MembersSection: FC = () => {
   if (acterLoading) return <LoadingSpinner />
   if (!acter) return null
 
-  const allFollowers = mapFollowersByType(acter)
-
-  const followers =
-    activeSelector === PEOPLE ? allFollowers.user : allFollowers.organisation
-
-  const validFollowers = followers?.filter((follower) =>
-    [ADMIN, MEMBER].includes(follower.role as ActerConnectionRole)
-  )
+  const validFollowers = getFollowersByType(acter, activeSelector)
 
   const handleSelectorChange = (selector) => setActiveSelector(selector)
 
-  const selectors = getActerJoinSelectors(acter)
+  const selectors = [PEOPLE, ACTERS]
 
   return (
     <Box className={classes.container}>
