+++ packages/lib/acter/map-followers-by-type.ts
@@ -1,6 +1,7 @@
-import { ActerTypes } from '../constants'
-
+import { ActerTypes, MemberType } from '@acter/lib/constants'
 import { Acter, ActerConnection } from '@acter/schema'
+
+const { PEOPLE, ACTERS } = MemberType
 
 export const mapFollowersByType = (
   acter: Acter
@@ -8,18 +9,10 @@ export const mapFollowersByType = (
   return (
     acter.Followers?.reduce((map, connection) => {
       const type = connection.Follower.ActerType?.name
-      const currentMap = map[type] || map['acters'] || []
-
-      if (type === ActerTypes.USER) {
-        return {
-          ...map,
-          ['user']: [...currentMap, connection],
-        }
-      } else if (type !== ActerTypes.USER) {
-        return {
-          ...map,
-          ['acters']: [...currentMap, connection],
-        }
+      const key = type === ActerTypes.USER ? PEOPLE : ACTERS
+      return {
+        ...map,
+        [key]: [...(map[key] || []), connection],
       }
     }, {}) || {}
   )
